Move MainWindow.java and DialogBox.java into package views
* Refactor JennyBot.java for the files moved.
* Refactor Main.java for the files moved.
* Refactor MainWindow.java for the files moved.
* Refactor MainWindow.fxml for the files moved.
* Modify JavaDoc for JennyBot::getResponse
* Modify access for JennyBot::getResponse
