+++ src/main/java/jenny/Main.java
@@ -1,12 +1,13 @@
 package jenny;
 
-import java.io.IOException;
-
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
+import jenny.views.MainWindow;
+
+import java.io.IOException;
 
 /**
  * A GUI for Duke using FXML.
