+++ src/main/java/jenny/views/DialogBox.java
@@ -1,7 +1,4 @@
-package jenny;
-
-import java.io.IOException;
-import java.util.Collections;
+package jenny.views;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -13,6 +10,9 @@ import javafx.scene.control.Label;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.HBox;
+
+import java.io.IOException;
+import java.util.Collections;
 
 /**
  * An example of a custom control using FXML.
