+++ src/main/java/jenny/views/MainWindow.java
@@ -1,4 +1,4 @@
-package jenny;
+package jenny.views;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
@@ -7,6 +7,7 @@ import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
+import jenny.JennyBot;
 
 import java.util.Objects;
